[Liverpool] Replace the #nav-link background image with CSS
This changes adds a class `.nav-link` which behaves in a very similar
way to the id `#nav-link`, but instead of using a background image that
needs to change everytime a new cobrand is added it uses CSS, which
means it will automatically respect the $primary colour that has been
set in the scss.

This could potentially be useful for FMS in general to prevent any
future issues with this link being the correct colour, but I notice that
some cobrands (Zurich) have a nav link which is a different colour to
the primary colour, something to think about.

https://github.com/mysociety/FixMyStreet-Commercial/issues/453